Factor out an impl declaration helper function
